Fixes
If you take out the enemy while Blocking it will get stuck. The enemy must be able to die while blocking too.

The Enemies are seen behind the objects. They must be seen in front of them. The Sorting Layer must be changed

The reference between AIController and the Stagger State was not properly set blocking the continuation of Behavior. Now it is properly set

The enemies could not pass from the Stagger state to the Stunned state. If you depleted the posture bar during stagger the enemy would be inmobile till dead.

The enemies can still be interrupted even when they are attacking. They must not be interrupted once the Attack Animation starts. Can only be interrupted if Staggered, Stunned or Dead.