Damage Penetration Percentage Added to the Weapons
The initial flow for this mechanic to work is complete. All the objects that use the Interface IDamagable need to use the scriptableObject Weapon, where the damage and the damagePenetrationPercentage is located for all the damaging objects.

Remains to define how it will be refactorized this objects since the Dagger is an Item as well as a Weapon. So this is kind of messy right now