Necessary Updates for the Player Parry to be 100% successful
Player  Guard and Player Health were refactorized to let Player Health what to do based on current State of the player.

This caused that the Interface IDamagable to know the AttackType to differentiate the Attack Type (Unblockable or not unblockable)

All entities and objects that used the IDamagable were updated for the previous change.